testapp/requirements.txt
-e ..
pytest==6.2.5
pytest-cov==2.12.1
pytest-django==4.4.0
black==20.8b1
mysqlclient==2.0.3
psycopg2-binary==2.9.2
